создает константы и разбивает код на модули
